Organize scripts dir.
