+++ frontend/scripts/icons/font_icons.json
